refactor: move local storage services to `local_storage_service/`
