+++ lib/local_storage_service/local_storage_service.dart
@@ -1,0 +1,4 @@
+export 'core/local_storage.dart';
+export 'core/local_storage_form.dart';
+export 'src/sign_in_local_storage_service.dart';
+export 'src/user_account_local_storage_service.dart';
