+++ lib/main.dart
@@ -6,12 +6,9 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:hydrated_bloc/hydrated_bloc.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:simple_notes_app/core/local_storage/local_storage.dart';
-import 'package:simple_notes_app/core/local_storage/local_storage_form.dart';
+import 'package:simple_notes_app/local_storage_service/local_storage_service.dart';
 import 'package:simple_notes_app/router/app_router.dart';
 import 'package:simple_notes_app/service/service.dart';
-import 'package:simple_notes_app/service/src/authentication/sign_in_local_storage_service.dart';
-import 'package:simple_notes_app/service/src/authentication/user_account_local_storage_service.dart';
 import 'package:simple_notes_app/service/src/sign_in_repository.dart';
 import 'package:simple_notes_app/service/src/user_account_repository.dart';
 import 'package:simple_notes_app/theme/app_theme.dart';
