+++ lib/service/src/sign_in_repository.dart
@@ -1,7 +1,7 @@
 import 'package:fpdart/fpdart.dart';
 import 'package:simple_notes_app/core/typedefs.dart';
+import 'package:simple_notes_app/local_storage_service/local_storage_service.dart';
 import 'package:simple_notes_app/models/models.dart';
-import 'package:simple_notes_app/service/src/authentication/sign_in_local_storage_service.dart';
 import 'package:simple_notes_app/service/src/authentication/sign_in_remote_service.dart';
 
 mixin ISignInRepository {
